requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
black==26.1.0
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
curl_cffi==0.14.0
davey==0.1.0
discord==2.3.2
discord-protos==0.0.2
discord.py==2.6.4
distro==1.9.0
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.61.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.5.0
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
selfcord.py @ git+https://github.com/dolfies/discord.py-self@060594e2874692190b513d1d399eb10154c2c2eb
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
